Conflicts:
	core/model/src/main/java/org/activityinfo/model/Model.gwt.xml
	core/model/src/main/java/org/activityinfo/model/date/CalendarUtils.java
	core/model/src/main/java/org/activityinfo/model/date/DateRange.java
	core/model/src/main/java/org/activityinfo/model/date/DayOfWeek.java
	core/model/src/main/java/org/activityinfo/model/date/EpiWeek.java
	core/model/src/main/java/org/activityinfo/model/form/FormClass.java
	core/model/src/main/java/org/activityinfo/model/form/FormField.java
	core/model/src/main/java/org/activityinfo/model/form/FormSection.java
	core/model/src/main/java/org/activityinfo/model/type/TypeRegistry.java
	core/model/src/test/java/org/activityinfo/model/date/CalendarUtilsTest.java
	i18n/src/main/java/org/activityinfo/i18n/shared/UiConstants.java
	server/src/main/java/org/activityinfo/legacy/shared/adapter/CriteriaAnalysis.java
	server/src/main/java/org/activityinfo/legacy/shared/adapter/projection/SiteProjector.java
	server/src/main/java/org/activityinfo/legacy/shared/impl/GetSchemaHandler.java
	server/src/main/java/org/activityinfo/legacy/shared/model/IndicatorDTO.java
	server/src/main/java/org/activityinfo/server/command/handler/GetFormClassHandler.java
	server/src/main/java/org/activityinfo/server/database/hibernate/entity/Activity.java
	server/src/main/java/org/activityinfo/server/database/hibernate/entity/LocationType.java
	server/src/main/java/org/activityinfo/server/endpoint/odk/FormSubmissionResource.java
	server/src/main/java/org/activityinfo/ui/client/component/form/SimpleFormPanel.java
	server/src/main/java/org/activityinfo/ui/client/component/formdesigner/FormDesigner.java
	server/src/main/java/org/activityinfo/ui/client/component/formdesigner/FormDesignerPanel.java
	server/src/main/java/org/activityinfo/ui/client/component/formdesigner/container/FieldPanel.java
	server/src/main/java/org/activityinfo/ui/client/component/formdesigner/container/FieldWidgetContainer.java
	server/src/main/java/org/activityinfo/ui/client/component/formdesigner/container/SectionWidgetContainer.java
	server/src/main/java/org/activityinfo/ui/client/component/table/InstanceTableDataLoader.java
	server/src/main/java/org/activityinfo/ui/client/page/config/design/DesignPresenter.java
	server/src/main/resources/org/activityinfo/database/changelog/db.changelog-master.xml
	server/src/main/resources/persistent.classes
	server/src/test/java/org/activityinfo/legacy/shared/adapter/ResourceLocatorAdaptorTest.java
	server/src/test/java/org/activityinfo/legacy/shared/adapter/SitePersisterTest.java
	server/src/test/java/org/activityinfo/ui/client/page/config/design/MockDesignTree.java

+++ server/src/main/java/org/activityinfo/server/command/handler/DeleteHandler.java
@@ -25,33 +25,46 @@ package org.activityinfo.server.command.handler;
 import com.google.inject.Inject;
 import org.activityinfo.legacy.shared.command.Delete;
 import org.activityinfo.legacy.shared.command.result.CommandResult;
-import org.activityinfo.server.database.hibernate.entity.*;
+import org.activityinfo.legacy.shared.exception.CommandException;
+import org.activityinfo.server.database.hibernate.entity.Deleteable;
+import org.activityinfo.server.database.hibernate.entity.ReallyDeleteable;
+import org.activityinfo.server.database.hibernate.entity.User;
+import org.activityinfo.server.database.hibernate.entity.UserDatabase;
 
 import javax.persistence.EntityManager;
-import java.util.Date;
+import java.util.logging.Logger;
 
 public class DeleteHandler implements CommandHandler<Delete> {
+    
+    private static final Logger LOGGER = Logger.getLogger(DeleteHandler.class.getName());
+    
     private EntityManager em;
+    private PermissionOracle permissionOracle;
 
     @Inject
-    public DeleteHandler(EntityManager em) {
+    public DeleteHandler(EntityManager em, PermissionOracle permissionOracle) {
         this.em = em;
+        this.permissionOracle = permissionOracle;
     }
 
     @Override
     public CommandResult execute(Delete cmd, User user) {
-        // TODO check permissions for delete!
-        // These handler should redirect to one of the Entity policy classes.
+
+        
         Class entityClass = entityClassForEntityName(cmd.getEntityName());
         Object entity = em.find(entityClass, cmd.getId());
+        
+        if(entity == null) {
+            throw new CommandException(String.format("%s with id %d does not exist.", 
+                    cmd.getEntityName(), cmd.getId()));
+        }
+
+        // Ensure that the user is authorized to perform deletion
+        permissionOracle.assertDeletionAuthorized(entity, user);
 
         if (entity instanceof Deleteable) {
             Deleteable deleteable = (Deleteable) entity;
             deleteable.delete();
-
-            if (entity instanceof Site) {
-                ((Site) entity).setDateEdited(new Date());
-            }
         }
 
         if (entity instanceof ReallyDeleteable) {
@@ -62,6 +75,7 @@ public class DeleteHandler implements CommandHandler<Delete> {
 
         return null;
     }
+
 
     private Class<Deleteable> entityClassForEntityName(String entityName) {
         try {
